requirements.txt
requests>=2.31.0
anthropic>=0.25.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
colorama>=0.4.6
flask>=3.0.0
